requirements.txt
chromadb==0.5.3
groq==0.7.0
python-dotenv
streamlit==1.34.0
pathlib
streamlit_extras